# api/docs/api.yml
openapi: '3.0.2'
info:
  title: API
  description: API description
  version: 0.0.0
servers:
  - url: https://contractus-25fea2a1cfb3.herokuapp.com
    description: Production server
  - url: http://localhost:8080
    description: Local server

paths:
  /upload:
    post:
      summary: Upload a transaction file
      tags:
        - "transactions"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - cookieAuth: []
  /transactions:
    get:
      summary: Return all transactions
      tags:
        - "transactions"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  total: 
                    type: integer
                    description: Number of transactions
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - cookieAuth: []
  /balance/affiliate:
    get:
      summary: Return balance for an affiliate
      tags:
        - transactions
      parameters:
        - name: name
          in: query
          description: Affiliate name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: number
                    description: Balance for affiliate
                  seller_name:
                    type: string
                    description: Affiliate name
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - cookieAuth: []
  /balance/producer:
    get:
      parameters:
        - name: name
          in: query
          description: Producer name
          required: true
          schema:
            type: string
      summary: Return balance for a producer
      tags:
        - transactions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: number
                    description: Balance for producer
                  seller_name:
                    type: string
                    description: Producer name
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - cookieAuth: []
  /token:
    get:
      summary: Return a token
      tags:
        - auth
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Token
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code error
        msg:
          type: string
          description: Error message
    Transaction:
      type: object
      properties:
        email: 
          type: string
          description: User email
        type:
          type: string
          description: Transaction type
        date:
          type: string
          format: date-time
          description: Transaction date
        product_description:
          type: string
          description:  Production description
        product_price:
          type: string
          description: Production price
        seller_name:
          type: string
          description: Seller name
        seller_type:
          type: string
          description: Seller type
          enum:
            - affiliate
            - producer
        action:
          type: string
          description: Transaction action
          enum:
            - venda produtor
            - venda afiliado
            - comissão afiliado
            - comissão produtor
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt